package all src files but main.py into zip
