--- UNKNOWN.egg-info/SOURCES.txt
@@ -1,6 +1,0 @@
-README
-pyproject.toml
-UNKNOWN.egg-info/PKG-INFO
-UNKNOWN.egg-info/SOURCES.txt
-UNKNOWN.egg-info/dependency_links.txt
-UNKNOWN.egg-info/top_level.txt
